The big bad push

+++ app/src/main/java/nl/teamone/projectholiday/api/responses/openweathermap/Wind.java
@@ -3,6 +3,7 @@ package nl.teamone.projectholiday.api.responses.openweathermap;
 public class Wind {
 
     public double speed;
-    public int deg;
+    public double deg;
     public double gust;
+
 }
